allow editing of feed/feed version import level
closes #22
