--- app/components/feed-eater-controls/template.hbs
@@ -1,2 +1,0 @@
-<input type="button" {{action 'enqueue' 1}} value="Enqueue Level 1" class="btn btn-danger" title="stops and routes">
-<input type="button" {{action 'enqueue' 2}} value="Enqueue Level 2" class="btn btn-danger" title="stops, routes, and schedules">
