+++ app/components/feed-versions-list/template.hbs
@@ -12,7 +12,9 @@
               this is the <strong style="color: green;">active feed version</strong>
             </small>
           {{/if}}
-          <div class="pull-right">{{feed-eater-controls feedVersion=feedVersion}}</div>
+          <div class="pull-right">
+            {{feed-version-controls feedVersion=feedVersion}}
+          </div>
         </h4>
       </div>
       <div class="panel-body">
@@ -30,6 +32,8 @@
             <dd>{{feedVersion.fetched_at}}</dd>
             <dt>Imported at</dt>
             <dd>{{feedVersion.imported_at}}</dd>
+            <dt>Import level</dt>
+            <dd>{{feedVersion.import_level}}</dd>
             <dt>Tags</dt>
             <dd>{{stringify-json feedVersion.tags}}</dd>
           </dl>
